backend/requirements.txt
fastapi
uvicorn
pytest
pytest-asyncio
pydantic
openai
elevenlabs
praat-parselmouth
numpy
python-multipart
python-dotenv
httpx
sqlmodel
passlib[bcrypt]
python-jose[cryptography]
